mixpanel/requirements.txt
apache-airflow==1.9.0
cryptography==2.4.2
google-api-core==1.5.2
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.7.0
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
pylint==2.2.2
pyjsparser==2.5.2
requests==2.20.1
